requirements.txt
git+https://github.com/metabrainz/brainzutils-python.git@v1.4.0
Fabric == 1.10.2
Flask-Testing == 0.4.2
Flask-SQLAlchemy==2.0
Flask-UUID == 0.2
Jinja2 == 2.8
SQLAlchemy==1.0.8
Werkzeug == 0.10.4
click == 4.1
coverage == 4.4.1
nose == 1.3.7
psycopg2 == 2.7.3.2
redis == 2.10.3
raven[flask] == 5.5.0
setproctitle == 1.1.9
six == 1.10.0
ujson==1.33
pytest==3.3.1
pytest-cov==2.5.1
unittest2==1.1.0
